+++ circuit/src/layered/shared_mem.rs
@@ -56,11 +56,11 @@ impl<C: GKRFieldConfig> SharedMemory for CircuitLayer<C> {
 impl<C: GKRFieldConfig> SharedMemory for Circuit<C> {
     fn bytes_size(&self) -> usize {
         self.layers.len().bytes_size()
-         + self
-            .layers
-            .iter()
-            .map(|layer| layer.bytes_size())
-            .sum::<usize>()
+            + self
+                .layers
+                .iter()
+                .map(|layer| layer.bytes_size())
+                .sum::<usize>()
             + self.expected_num_output_zeros.bytes_size()
     }
