+++ config/mpi_config/src/shared_mem.rs
@@ -58,8 +58,7 @@ impl SharedMemory for u8 {
 
 impl<T: Copy> SharedMemory for Vec<T> {
     fn bytes_size(&self) -> usize {
-        self.len().bytes_size()
-         + self.len() * std::mem::size_of::<T>()
+        self.len().bytes_size() + self.len() * std::mem::size_of::<T>()
     }
 
     fn to_memory(&self, ptr: &mut *mut u8) {
